packages/ai-analysis/requirements.txt
flask==3.0.0
pymongo==4.6.1
pika==1.3.2
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.26.2
